Move `makeEmpty` from `Storage` to `Builder` for the numerics.
