+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/BigDecimalBuilder.java
@@ -8,6 +8,12 @@ import org.enso.table.error.ValueTypeMismatchException;
 
 /** A builder for BigDecimal columns. */
 public final class BigDecimalBuilder extends TypedBuilder<BigDecimal> {
+  /** Creates a new empty BigDecimal storage with the specified size. */
+  public static Storage<BigDecimal> makeEmpty(long size) {
+    int intSize = Builder.checkSize(size);
+    return new BigDecimalStorage(new BigDecimal[intSize]);
+  }
+
   BigDecimalBuilder(int size) {
     super(BigDecimalType.INSTANCE, new BigDecimal[size]);
   }
